Move some IOD functions to core
